+++ src/components/ControlPanel.jsx
@@ -1,17 +1,11 @@
-/* eslint-disable array-callback-return */
-/* eslint-disable no-unused-vars */
-/* eslint-disable prefer-const */
 import React, { useContext, useState } from 'react';
 import styled from 'styled-components';
-import { MoneyContext, CoinContext, EventLogContext } from 'components/App';
+import { MoneyContext } from 'components/App';
 import ModifiableInput from 'components/ModifiableInput';
 import EventLog from 'components/EventLog';
-import { copyObject } from 'utils';
 
 function ControlPanel() {
-  const { curMoney, setMoney } = useContext(MoneyContext);
-  const { coins, setCoins } = useContext(CoinContext);
-  const { eventLog, setEventLog } = useContext(EventLogContext);
+  const { curMoney, handleReturn } = useContext(MoneyContext);
   const [isInputMode, setInputMode] = useState(false);
   const UnmodifiableInput = React.useCallback(
     () => getUnmodifiableInput({ value: curMoney, handler: handleInputMode }),
@@ -23,35 +17,12 @@ function ControlPanel() {
       <Row>
         {isInputMode ? <ModifiableInput setInputMode={setInputMode} /> : <UnmodifiableInput />}
       </Row>
-      <button type="button" onClick={handleReturnBtn}>
+      <button type="button" onClick={handleReturn}>
         반환하기
       </button>
       <EventLog />
     </Wrap>
   );
-
-  function handleReturnBtn() {
-    const newCoins = updateCoins();
-
-    setCoins(newCoins);
-    setEventLog([...eventLog, { type: 'RETURN', value: curMoney }]);
-    setMoney(0);
-
-    function updateCoins() {
-      let leftMoney = curMoney;
-      const copiedCoins = coins.map(copyObject);
-      const updatedCoins = copiedCoins.map((coin) => {
-        const quotient = Math.floor(leftMoney / coin.AMOUNT);
-        if (quotient) {
-          leftMoney -= coin.AMOUNT * quotient;
-          return { ...coin, CNT: coin.CNT + quotient };
-        }
-        return coin;
-      });
-
-      return updatedCoins;
-    }
-  }
 
   function handleInputMode() {
     if (isInputMode) {
